refactor: introduce BaseAop
