+++ src/main/java/com/meoguri/linkocean/configuration/security/jwt/SecurityUser.java
@@ -40,12 +40,4 @@ public final class SecurityUser extends User {
 		return profileId != null;
 	}
 
-	@Override
-	public String toString() {
-		return new StringBuilder()
-			.append("SecurityUser{")
-			.append("id=").append(id).append(", ")
-			.append("profileId=").append(profileId).append("}")
-			.toString();
-	}
 }
